backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
yfinance>=1.3.0
pandas>=2.0.0
langchain==0.2.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
groq==0.9.0
httpx==0.27.0
